Finished the Fees tab
